requirements.txt
pymongo[srv]==3.12
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1